update type as per the latest mobile token implementation
