+++ src/login/mobile-token-onboarding/MobileToken.tsx
@@ -1,36 +1,35 @@
 import React from 'react';
 import {View} from 'react-native';
-import {TCardAsMobileToken} from '@atb/login/mobile-token-onboarding/TCardAsMobileToken';
-import {PhoneAsMobileToken} from '@atb/login/mobile-token-onboarding/PhoneAsMobileToken';
+import {TCardAsToken} from '@atb/login/mobile-token-onboarding/TCardAsToken';
+import {MobileAsToken} from '@atb/login/mobile-token-onboarding/MobileAsToken';
 import {StyleSheet} from '@atb/theme';
 import {StaticColorByType} from '@atb/theme/colors';
 import {NoMobileToken} from '@atb/login/mobile-token-onboarding/NoMobileToken';
-import {TravelToken} from '@atb/mobile-token/types';
+import {RemoteToken} from '@atb/mobile-token/types';
+import {isMobileToken, isTravelCardToken} from '@atb/mobile-token/utils';
 
 const themeColor: StaticColorByType<'background'> = 'background_accent_0';
 
-const MobileToken = ({inspectableToken}: {inspectableToken: TravelToken}) => {
+const MobileToken = ({inspectableToken}: {inspectableToken: RemoteToken}) => {
   const styles = useStyles();
-  switch (inspectableToken?.type) {
-    case 'travelCard':
-      return (
-        <View style={styles.container}>
-          <TCardAsMobileToken inspectableToken={inspectableToken} />
-        </View>
-      );
-    case 'mobile':
-      return (
-        <View style={styles.container}>
-          <PhoneAsMobileToken inspectableToken={inspectableToken} />
-        </View>
-      );
-    default:
-      return (
-        <View style={styles.container}>
-          <NoMobileToken />
-        </View>
-      );
+  if (isMobileToken(inspectableToken)) {
+    return (
+      <View style={styles.container}>
+        <MobileAsToken inspectableToken={inspectableToken} />
+      </View>
+    );
+  } else if (isTravelCardToken(inspectableToken)) {
+    return (
+      <View style={styles.container}>
+        <TCardAsToken inspectableToken={inspectableToken} />
+      </View>
+    );
   }
+  return (
+    <View style={styles.container}>
+      <NoMobileToken />
+    </View>
+  );
 };
 
 const useStyles = StyleSheet.createThemeHook((theme) => ({
