+++ src/login/mobile-token-onboarding/TCardAsToken.tsx
@@ -7,39 +7,37 @@ import Button from '@atb/components/button';
 import {StyleSheet, useTheme} from '@atb/theme';
 import React from 'react';
 import {flatStaticColors, StaticColorByType} from '@atb/theme/colors';
-import {ThemedTokenPhone} from '@atb/theme/ThemedAssets';
-import useFocusOnLoad from '@atb/utils/use-focus-on-load';
-import {settingToRouteName} from '@atb/utils/navigation';
+import {ThemedTokenTravelCard} from '@atb/theme/ThemedAssets';
 import {usePreferenceItems} from '@atb/preferences';
+import {settingToRouteName} from '@atb/utils/navigation';
 import {TravelDeviceTitle} from '@atb/travel-token-box';
-import {TravelToken} from '@atb/mobile-token/types';
+import {RemoteToken} from '@atb/mobile-token/types';
 import {useAppState} from '@atb/AppContext';
 
 const themeColor: StaticColorByType<'background'> = 'background_accent_0';
 
-export function PhoneAsMobileToken({
+export function TCardAsToken({
   inspectableToken,
 }: {
-  inspectableToken: TravelToken;
+  inspectableToken: RemoteToken;
 }): JSX.Element {
   const styles = useThemeStyles();
   const {t} = useTranslation();
   const navigation = useNavigation();
   const {themeName} = useTheme();
-  const focusRef = useFocusOnLoad();
   const {startScreen} = usePreferenceItems();
   const {completeMobileTokenOnboarding} = useAppState();
   return (
     <View style={styles.container}>
       <View style={styles.mainView}>
-        <View accessible={true} ref={focusRef}>
+        <View>
           <ThemeText
             type="body__primary--jumbo"
             style={[styles.alignCenter, styles.marginVertical]}
             color={themeColor}
             isMarkdown={true}
           >
-            {t(MobileTokenOnboardingTexts.phone.heading)}
+            {t(MobileTokenOnboardingTexts.tCard.heading)}
           </ThemeText>
         </View>
         <View
@@ -53,9 +51,9 @@ export function PhoneAsMobileToken({
         >
           <TravelDeviceTitle inspectableToken={inspectableToken} />
           <View style={styles.flex}>
-            <ThemedTokenPhone />
+            <ThemedTokenTravelCard />
             <ThemeText color={themeColor} style={styles.reminderText}>
-              {t(MobileTokenOnboardingTexts.phone.reminder)}
+              {t(MobileTokenOnboardingTexts.tCard.reminder)}
             </ThemeText>
           </View>
         </View>
@@ -64,46 +62,50 @@ export function PhoneAsMobileToken({
           color={themeColor}
           isMarkdown={true}
         >
-          {t(MobileTokenOnboardingTexts.phone.description)}
+          {t(MobileTokenOnboardingTexts.tCard.description)}
         </ThemeText>
       </View>
+
       <View style={styles.bottomView}>
-          <Button
-            onPress={() => {
-              completeMobileTokenOnboarding();
-              navigation.navigate(settingToRouteName(startScreen));
-            }}
-            text={t(MobileTokenOnboardingTexts.ok)}
-            testID="nextButton"
-            style={styles.nextButton}
-          />
-          <Button
-            interactiveColor="interactive_1"
-            mode="secondary"
-            onPress={() => {
-              navigation.navigate('SelectTravelToken');
-            }}
-            text={t(MobileTokenOnboardingTexts.phone.button)}
-            testID="switchButton"
-          />
+        <Button
+          interactiveColor="interactive_0"
+          onPress={() => {
+            completeMobileTokenOnboarding();
+            navigation.navigate(settingToRouteName(startScreen));
+          }}
+          text={t(MobileTokenOnboardingTexts.ok)}
+          testID="nextButton"
+          style={styles.nextButton}
+        />
+        <Button
+          interactiveColor="interactive_1"
+          mode="secondary"
+          onPress={() => {
+            navigation.navigate('SelectTravelToken');
+          }}
+          text={t(MobileTokenOnboardingTexts.tCard.button)}
+          testID="switchButton"
+        />
       </View>
     </View>
   );
 }
-
 const useThemeStyles = StyleSheet.createThemeHook((theme) => ({
   container: {
     paddingHorizontal: theme.spacings.xLarge,
+  },
+  alignCenter: {
+    textAlign: 'center',
   },
   mainView: {
     marginTop: 164,
     height: 430,
   },
-  alignCenter: {
-    textAlign: 'center',
-  },
   marginVertical: {
     marginTop: 44,
+  },
+  cardNo: {
+    flexDirection: 'row',
   },
   description: {
     textAlign: 'center',
@@ -112,39 +114,19 @@ const useThemeStyles = StyleSheet.createThemeHook((theme) => ({
   },
   flex: {
     flexDirection: 'row',
+    width: 327,
   },
   reminderText: {
-    marginHorizontal: theme.spacings.medium,
-    marginTop: theme.spacings.large,
-    marginRight: 80,
+    marginLeft: theme.spacings.medium,
+    marginTop: theme.spacings.medium,
+    marginRight: 160,
   },
   mobileTokenContainer: {
     backgroundColor: theme.static.background.background_accent_2.background,
     height: 178,
     marginVertical: theme.spacings.xLarge,
     borderRadius: theme.border.radius.regular,
-    padding: 28,
-  },
-  phoneInfoBox: {
-    marginVertical: theme.spacings.xLarge,
-    alignSelf: 'center',
-    padding: theme.spacings.xLarge,
-    borderRadius: theme.border.radius.circle,
-    minHeight: 300,
-    minWidth: 200,
-  },
-  phoneLine: {
-    width: theme.spacings.xLarge * 2,
-    borderRadius: theme.border.radius.regular,
-    height: theme.spacings.small,
-    alignSelf: 'center',
-    marginTop: theme.spacings.small,
-    marginBottom: theme.spacings.small + theme.spacings.xLarge,
-  },
-  phoneInfoBoxInner: {
-    borderRadius: theme.border.radius.regular,
-    padding: theme.spacings.large,
-    alignSelf: 'center',
+    padding: 24,
   },
   bottomView: {
     marginTop: 35.5,
